middleware

+++ routes/web.php
@@ -28,14 +28,8 @@
     Auth::routes();
 
     Route::get('/home', [App\Http\Controllers\HomeController::class, 'index'])->name('home');
-    Route::get('/dashboard', function () {
-        return view('dashbord');
-    });
+    
 
-    // Route::get('/hotel', function () {
-    //     return view('hotel');
-    // });
-   
     Route::get('/hotelHome', function () {
         return view('hotelHome');
     });
@@ -46,33 +40,26 @@
          
     Route::get('/home', [HomeController::class, 'index'])->name('home');
 
-    Route::resource('articles', ArticleController::class);
-    Route::resource('comments', CommentController::class);
-    Route::resource('likes', LikeController::class);
-    Route::resource('hotel', HotelController::class);
-    Route::resource('transports', TransportController::class);
-    Route::resource('blogDashboard', ArticleAdminController::class);
+    Route::resource('articles', ArticleController::class)->middleware('auth');
+    Route::resource('comments', CommentController::class)->middleware('auth');
+    Route::resource('likes', LikeController::class)->middleware('auth');
+    Route::resource('hotel', HotelController::class)->middleware('auth');
+    Route::resource('transports', TransportController::class)->middleware('auth');
+    Route::resource('blogDashboard', ArticleAdminController::class)->middleware('auth');
   
-    Route::get('/noha',[StatistiquesController::class ,'count']);
+    Route::get('/dashboard',[StatistiquesController::class ,'count'])->middleware('auth');
+
     Route::get('/hotels',[HotelController::class ,'indexClient'])->name('hotels');
 
     Route::get('/transportClient',[TransportController::class ,'indexClient'])->name('transportClient');
     Route::resource('voyageDashboard', VoyageController::class);
-    Route::get('/voyageOrganise',[VoyageController::class ,'showHome'])->name('voyageOrganise');
+    Route::get('/voyageOrganise',[VoyageController::class ,'showHome'])->name('voyageOrganise')->middleware('auth');
     Route::resource('reservation', ReservationController::class);
-    Route::get('/indexHome',[HotelController::class ,'indexHome'])->name('indexHome');
-    // patsh::resource('profile', UserController::class);
-    // Route::post('/profile/update', 'App\Http\Controllers\UserController@update')->name('profile.update');
-    // Route::get('password/reset', ForgotPasswordController::class,'howLinkRequestForm')->name('password.request');
-
-    // Route::post('password/email', ForgotPasswordController::class ,'sendResetLinkEmail')->name('password.email');
-    
-    // Route::get('password/reset/{token}', ResetPasswordController::class , 'showResetForm')->name('password.reset');
-    
-    // Route::post('password/reset', ResetPasswordController::class ,'reset')->name('password.update');
-    Route::get('/forgot-password', 'App\Http\Controllers\Auth\ForgotPasswordController@showLinkRequestForm');
+    Route::get('/indexHome',[HotelController::class ,'indexHome'])->name('indexHome')->middleware('auth');
+   
+        Route::get('/forgot-password', 'App\Http\Controllers\Auth\ForgotPasswordController@showLinkRequestForm')->middleware('auth');
     Route::prefix('profile')->group(function () {
-        Route::put('/{id}', [UserController::class, 'update'])->name('profile.update');
-        Route::delete('/{id}', [UserController::class, 'delete'])->name('profile.delete');
+        Route::put('/{id}', [UserController::class, 'update'])->name('profile.update')->middleware('auth');
+        Route::delete('/{id}', [UserController::class, 'delete'])->name('profile.delete')->middleware('auth');
     });
